added current git commit id to history on every itaSuper constructor
this will be the case on new itaAudios, or loaded ones. This will make itaAudios more reproducable in terms of toolbox version

the git hash is generated via system call and will return the current id of the master branch OR (in case of a different branch) the common ancestor of the current branch with the master branch.
this is done, in case work is done on a private branch.

Example History:
- Create Git hash
- ita_time_window
- Load Git hash
- ita_crop

etc